+++ src/world.hpp
@@ -285,9 +285,9 @@ public:
         AFK_Computer *computer,
         AFK_ThreadAllocation& threadAlloc,
         float _maxDistance,
-        unsigned int worldCacheSize, /* in bytes */
-        unsigned int tileCacheSize, /* also in bytes */
-        unsigned int shapeCacheSize, /* likewise */
+        size_t worldCacheSize, /* in bytes */
+        size_t tileCacheSize, /* also in bytes */
+        size_t shapeCacheSize, /* likewise */
         AFK_RNG *setupRng);
     virtual ~AFK_World();
